+++ Classes/Domain/Validator/AbstractValidator.php
@@ -25,6 +25,7 @@ namespace Portrino\PxValidation\Domain\Validator;
  *
  *  This copyright notice MUST APPEAR in all copies of the script!
  ***************************************************************/
+
 use Exception;
 use TYPO3\CMS\Extbase\Configuration\ConfigurationManager;
 use TYPO3\CMS\Extbase\Persistence\ObjectStorage;