+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 1.2.0 - 2019-01-08
 ------------------
 * [TASK] updates dependencies for TYPO3 9.5
+* [TASK] reworks validator loading with reflection
 
 1.1.2 - 2016-07-13
 ------------------